+++ oci_cli/image/import_.py
@@ -20,12 +20,12 @@ import tempfile
 import shutil
 import os
 from urllib.request import urlopen
-from urllib.parse import urlparse
 from oci_spec.image.v1 import ImageConfig
 from oci_spec.runtime.v1 import Spec
 from oci_api import OCIError
-from oci_api.image import Distribution, ImageExistsException
-
+from oci_api.util.file import untar
+from oci_api.image import Distribution
+from oci_api.graph import Driver
 log = logging.getLogger(__name__)
 
 class Import:
@@ -44,53 +44,57 @@ class Import:
         parser.add_argument('file',
             metavar='file|URL|-',
             help='Name of the file or URL to import, or "-" for the standard input')
-        parser.add_argument('image',
+        parser.add_argument('tag',
             metavar='REPOSITORY[:TAG]',
             nargs='?',
             help='Name of the repository to import to')
   
     def __init__(self, options):
-        image_name = options.image
-        log.debug('Start importing (%s) to (%s)', options.file, image_name)
+        log.debug('Start importing (%s)' % options.file)
+        layer = None
+        with tempfile.TemporaryDirectory() as temp_dir_name:
+            if options.file == '-':
+                input_file = os.fdopen(sys.stdin.fileno(), 'rb')
+            else:
+                try:
+                    input_file = urlopen(options.file)
+                except ValueError:
+                    input_file = open(options.file, 'rb')
+            rootfs_tar_path = pathlib.Path(temp_dir_name, 'rootfs.tar')
+            with rootfs_tar_path.open('wb') as output_file:
+                log.debug('Start copying (%s) to (%s)', options.file, str(rootfs_tar_path))
+                shutil.copyfileobj(input_file, output_file)
+                log.debug('Finish copying (%s) to (%s)', options.file, str(rootfs_tar_path))
+            if rootfs_tar_path.is_file():
+                filesystem = Driver().create_filesystem()
+                untar(filesystem.path, tar_file=input_file)
+                layer = Driver().create_layer(filesystem)
+        if layer is None:
+            log.error('Could not create layer')        
+            exit(-1)
         try:
-            with tempfile.TemporaryDirectory() as temp_dir_name:
-                if options.file == '-':
-                    input_file = os.fdopen(sys.stdin.fileno(), 'rb')
-                else:
-                    try:
-                        input_file = urlopen(options.file)
-                    except ValueError:
-                        input_file = open(options.file, 'rb')
-                rootfs_tar_path = pathlib.Path(temp_dir_name, 'rootfs.tar')
-                with rootfs_tar_path.open('wb') as output_file:
-                    log.debug('Start copying (%s) to (%s)', options.file, str(rootfs_tar_path))
-                    shutil.copyfileobj(input_file, output_file)
-                    log.debug('Finish copying (%s) to (%s)', options.file, str(rootfs_tar_path))
-                if rootfs_tar_path.is_file():
-                    distribution = Distribution()
-                    environment = None
-                    command = None
-                    working_dir = None
-                    if options.runc_config is not None:
-                        config_file_path = pathlib.Path(options.runc_config)
-                        if not config_file_path.is_file():
-                            OCIError('Runc config file (%s) does not exist' % str(config_file_path))
-                        spec = Spec.from_file(config_file_path)
-                        process = spec.get('Process')
-                        command = process.get('Args')
-                        environment = process.get('Env')
-                        working_dir = process.get('Cwd')
-                    image_config = ImageConfig(
-                        env=environment,
-                        cmd=command,
-                        working_dir=working_dir
-                    )
-                    image = distribution.import_image(image_name, rootfs_tar_path, image_config)
-        except ImageExistsException:
-            log.error('Image (%s) already exists' % image_name)
+            distribution = Distribution()
+            history = '/bin/sh -c #(nop) IMPORTED file:%s in / ' % options.file
+            image = Distribution().create_image(layer=layer, history=history)
+            if options.runc_config is not None:
+                config_file_path = pathlib.Path(options.runc_config)
+                if not config_file_path.is_file():
+                    OCIError('Runc config file (%s) does not exist' % str(config_file_path))
+                spec = Spec.from_file(config_file_path)
+                process = spec.get('Process')
+                command = process.get('Args')
+                if command is not None:
+                    image.set_command(command)
+                environment = process.get('Env')
+                if environment is not None:
+                    image.set_environment(environment)
+                working_dir = process.get('Cwd')
+                if working_dir is not None:
+                    image.set_working_dir(working_dir)
+                    if options.tag is not None:
+                        for tag in options.tag:
+                            Distribution().add_tag(self.image, tag)
+        except Exception as e:
+            log.error(e.args[0])
             exit(-1)
-        except:
-            raise e
-            log.error('Could not remove image (%s)' % image_name)
-            exit(-1)
-        log.debug('Finish importing (%s) to (%s)', options.file, image_name)
+        log.debug('Finish importing (%s)' % options.file)